+++ src/profile/__init__.py
@@ -4,6 +4,7 @@ User Profile Package
 This package manages user profiles, risk assessment, and investment goals.
 """
 
-from . import risk_assessment
-from . import profile_manager
-from . import goals_tracking 
+# TODO: Uncomment when profile modules are implemented
+# from . import risk_assessment
+# from . import profile_manager
+# from . import goals_tracking